Nunit-Basics/ConsoleAppSummator/Summator.cs
namespace ConsoleAppSummator
{
    public static class Summator
    {
        public static double Sum(double[] arr)
        {

            double sum = 0;

            for (int i = 0; i < arr.Length; i++)
            {
                sum += arr[i];
            }

            return sum;
        }

        public static double Average(double[] arr)
        {
            double sum = 0;

            for (int i = 0; i < arr.Length; i++)
            {
                sum += arr[i];
            }

            return sum / arr.Length;
        }

    }
}